ftp and script update

+++ srcs/requirements/bonus/adminer/tools/adminer.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 
 mkdir -p /run/php/
-
-wget -O adminer.php https://github.com/vrana/adminer/releases/download/v4.8.1/adminer-4.8.1.php
-
 mkdir -p /var/www/html
 
-chmod 777 ./adminer.php
-mv ./adminer.php /var/www/html/index.php
+if [ ! -f adminer.php ]; then
+    wget -O adminer.php https://github.com/vrana/adminer/releases/download/v4.8.1/adminer-4.8.1.php
+    chmod 777 ./adminer.php
+    cp ./adminer.php /var/www/html/index.php
+fi
+
 chown -R www-data:www-data /var/www/html
 
 echo "Adminer is ready!"
